% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcula.nota.R
\name{calc.nota}
\alias{calc.nota}
\title{Calcula nota do Enem}
\usage{
calc.nota(resps, codigo = NULL, lingua = NULL)
}
\arguments{
\item{resps}{vetor de respostas de um ou mais sujeitos. O
vetor deve ser no mesmo formato da variável TX_RESPOSTA
dos microdados do Enem. É importante destacar que desde 2010 nos
cadernos de Linguagens e Códigos existem cinco itens de língua
inglesa e cinco itens de língua espanhola, por isso em algumas edições
o vetor de
respostas nessa área possui 50 caracteres. A pessoa responde
somente uma dessas línguas e nos microdados as respostas aos cinco
itens da outra língua são marcados com "9". A função \code{calcula.nota}
automaticamente transforma essa resposta "9" em \code{NA} e esses
itens não considerados para o cálculo.}

\item{codigo}{código da prova disponível no dicionário dos
microdados. Essa informação também está disponível no
objeto \code{dic.cad} deste pacote}

\item{lingua}{vetor com indicação da língua estrangeira escolhida
pela pessoa. Essa informação corresponde à variável \code{TP_LINGUA}
nos microdados (\code{0} para inglês, \code{1} para espanhol).}
}
\value{
as notas na escala oficial do Enem
}
\description{
Calcula a nota de um sujeito a escala do Enem (500, 100)
}
\details{
A nota é calculada pelo método expected a
posteriori (EAP), com 40 pontos de quadratura de -4 a 4.
A média da distribuição priori é 0 e o desvio padrão, 1. No
Enem as notas estão em uma escala com média 500 e desvio padrão
100. A referência dessa escala são os concluintes regulares de
escola pública do Enem 2009. Ou seja, a média desses alunos
no Enem 2009 foi 500 e o desvio padrão, 100.

Os parâmetros dos itens estão divulgados em uma escala com
média 0 e desvio padrão 1. A referência dessa escala é a
amostra utilizada na primeira calibração dos itens do Enem,
em 2009. Para posicionar os parâmetros na a escala oficial
do Enem, aplicamos as seguintes equações de
transformação:
\deqn{a_{enem} = \frac{a_{01}}{s}}
\deqn{b_{enem} = b_{01}*s+m}
\deqn{c_{enem} = c_{01}}
Onde \eqn{a_{enem}}, \eqn{b_{enem}} e \eqn{c_{enem}} são os
parâmetros dos itens na escala oficial do Enem,
\eqn{a_{01}}, \eqn{b_{01}} e \eqn{c_{01}} são os parâmetros
dos itens divulados nos microdados, e \code{k} e \code{d} são as
constantes de transformação da escala dos parâmetros divulgados
para a escala oficial. Essas constantes estão disponibilizadas
no objeto \code{constantes} deste pacote.
}
\examples{
# importar primeiros 100 casos dos microdados de 2019
micro <- data.table::fread('MICRODADOS_ENEM_2019.csv', nrows = 100)

# selecionar quatro casos do cadernos 511 (LC, primeira aplicação)
resp <- micro$TX_RESPOSTAS_LC[c(84, 97:99)]

# calcular a nota
nota <- calc.nota(resp, codigo = 511)
nota

# comparar com a nota oficial
micro$NU_NOTA_LC[c(84, 97:99)]

# calcular a nota de um sujeito que não está nos microdados
# vetor de resposta de uma pessoa fictícia que respondeu o caderno de Linguagens e Códigos em 2019
resp <- c('99999BBDABBDBAADCBABBADAACBDDDDEACACBCACAABBBECBEC')

# calcular a nota
nota <- calc.nota(resp, codigo = 511)
nota
}
