% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download_microdados}
\alias{download_microdados}
\alias{download_microdados_descompactado}
\title{Download dos microdados do ENEM}
\usage{
download_microdados(path, year)

download_microdados_descompactado(path, year)
}
\arguments{
\item{path}{Caminho para salvar o arquivo que será baixado.}

\item{year}{Ano para o qual queremos fazer o download.}
}
\value{
Retorna o próprio \code{path} de forma invisível.
}
\description{
Faz o download do arquivo .zip dos microdados do ENEM.
}
\section{Functions}{
\itemize{
\item \code{download_microdados_descompactado()}: Baixa e descompacta os arquivos

}}
\section{Organização dos arquivos}{
Os arquivos baixados incluem:
\itemize{
\item Dados:
\item Itens da prova
\item Microdados
\item Dicionário de dados
\item Inputs (códigos para leitura dos dados)
\item PDFS de provas e gabaritos
}
}

\section{Disponibilização dos dados}{
Os dados são disponibilizados pelo INEP, na página
de \href{https://www.gov.br/inep/pt-br/acesso-a-informacao/dados-abertos/microdados/enem}{dados abertos}.
}

\examples{
\dontrun{
tmp <- tempfile(fileext = ".zip")
download_microdados(tmp, 2023)
}

}
