% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abre.resp.R
\name{abre.resp}
\alias{abre.resp}
\title{Abre o vetor de resposta}
\usage{
abre.resp(unico)
}
\arguments{
\item{unico}{vetor único de respostas de um ou mais sujeitos no
formato da variável TX_RESPOSTAS dos microdados do Enem}
}
\value{
Um objeto de classe \code{matrix} em que as colunas são os itens e as
linhas são os sujeitos
}
\description{
Abre o vetor de resposta da variável TX_RESPOSTAS dos microdados
}
\examples{

# importar primeiros 100 casos dos microdados de 2019 (é preciso direcionar para a pasta onde o arquivo se localiza)
micro <- data.table::fread('MICRODADOS_ENEM_2019.csv', nrows = 100)

resp <- abre.resp(micro$TX_RESPOSTAS_LC)
}
