% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod.caderno.R
\name{mod.caderno}
\alias{mod.caderno}
\title{Modelo mirt do caderno}
\usage{
mod.caderno(codigo = NULL)
}
\arguments{
\item{codigo}{código da prova disponível no dicionário dos
microdados. Essa informação também está disponível no
objeto \code{dic.cad} deste pacote}
}
\value{
Um objeto \code{mirt} com os
parâmetros dos itens. Para
informações sobre esse objeto, consulte a documentação do
pacote \code{mirt}.
}
\description{
Gera um modelo do pacote \code{mirt} para um caderno do Enem
}
\examples{

mod <- mod.caderno(511)

# parâmetros dos itens
mirt::coef(mod, IRTpars = TRUE, simplify = TRUE)

# gráfico com informação do item 1
item <- mirt::extract.item(mod, 1)
theta <- matrix(seq(-4,4, by = .1))
info <- mirt::iteminfo(item, theta)
plot(theta, info, type = 'l', main = 'Informação do item 1')

}
